Update WSEvents.cpp
/root/obs-websocket/WSEvents.cpp: In function 'const char* ns_to_timestamp(uint64_t)':
/root/obs-websocket/WSEvents.cpp:54:53: warning: format '%d' expects argument of type 'int', but argument 3 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
         hours_part, minutes_part, secs_part, ms_part);
                                                     ^
/root/obs-websocket/WSEvents.cpp:54:53: warning: format '%d' expects argument of type 'int', but argument 4 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
/root/obs-websocket/WSEvents.cpp:54:53: warning: format '%d' expects argument of type 'int', but argument 5 has type 'uint64_t {aka long unsigned int}' [-Wformat=]
/root/obs-websocket/WSEvents.cpp:54:53: warning: format '%d' expects argument of type 'int', but argument 6 has type 'uint64_t {aka long unsigned int}' [-Wformat=]

+++ WSEvents.cpp
@@ -51,7 +51,7 @@ const char* ns_to_timestamp(uint64_t ns) {
 
     char* ts = (char*)bmalloc(64);
     sprintf(ts, "%02d:%02d:%02d.%03d",
-        hours_part, minutes_part, secs_part, ms_part);
+        (int)hours_part, (int)minutes_part, (int)secs_part, (int)ms_part);
 
     return ts;
 }
